minor fixes for changes in type
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@200 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/SpecifyDBConverter.java
@@ -1263,7 +1263,7 @@ public class SpecifyDBConverter
                 BasicSQLUtils.deleteAllRecordsFromTable("collectionobjdef");
                 
                 DataType          dataType  = conversion.createDataTypes("Animal");        
-                User              user      = conversion.createNewUser("rods", "rods", 0);
+                User              user      = conversion.createNewUser("rods", "rods", (short)0);
                 
                 Criteria criteria = HibernateUtil.getCurrentSession().createCriteria(CatalogSeries.class);
                 criteria.add(Expression.eq("catalogSeriesId", new Integer(0)));
